refactor: import 경로 일부 수정 및 결과화면 새로고침 대응
