+++ src/pages/EditPage/EditPage.js
@@ -3,8 +3,8 @@ import { useRecoilValue, useSetRecoilState } from "recoil";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
 
-import SubPageLayout from "../SubPageLayout/SubPageLayout";
-import LayoutPreview from "../../components/LayoutPreview/LayoutPreview";
+import SubPageLayout from "../SubPageLayout";
+import LayoutPreview from "../../components/LayoutPreview";
 
 import { generatedCodeState, imageSizeState } from "../../recoil/store";
 import fitToMaxCanvasSize from "../../utils/fitToMaxCanvasSize";
